+++ ddtrace/_trace/tracer.py
@@ -15,8 +15,7 @@ from typing import Tuple
 from typing import TypeVar
 from typing import Union
 
-from ddtrace import _hooks
-from ddtrace import config
+from ddtrace._hooks import Hooks
 from ddtrace._trace.context import Context
 from ddtrace._trace.processor import SpanAggregator
 from ddtrace._trace.processor import SpanProcessor
@@ -64,6 +63,7 @@ from ddtrace.internal.writer import LogWriter
 from ddtrace.internal.writer import TraceWriter
 from ddtrace.settings._agent import config as agent_config
 from ddtrace.settings._config import Config
+from ddtrace.settings._config import config
 from ddtrace.settings.asm import config as asm_config
 from ddtrace.settings.peer_service import _ps_config
 
@@ -273,7 +273,7 @@ class Tracer(object):
             self.data_streams_processor = DataStreamsProcessor(self._agent_url)
             register_on_exit_signal(self._atexit)
 
-        self._hooks = _hooks.Hooks()
+        self._hooks = Hooks()
         atexit.register(self._atexit)
         forksafe.register_before_fork(self._sample_before_fork)
         forksafe.register(self._child_after_fork)
