+++ ddtrace/contrib/internal/trace_utils.py
@@ -21,7 +21,6 @@ from urllib import parse
 
 import wrapt
 
-from ddtrace import config
 from ddtrace.ext import http
 from ddtrace.ext import net
 from ddtrace.ext import user
@@ -36,6 +35,7 @@ from ddtrace.internal.utils.http import redact_url
 from ddtrace.internal.utils.http import strip_query_string
 import ddtrace.internal.utils.wrappers
 from ddtrace.propagation.http import HTTPPropagator
+from ddtrace.settings._config import config
 from ddtrace.settings.asm import config as asm_config
 from ddtrace.trace import Pin
