+++ ddtrace/settings/_config.py
@@ -940,3 +940,6 @@ class Config(object):
 
     def _lower(self, value):
         return value.lower()
+
+
+config = Config()
